+++ src/routes/cadastroRouter.js
@@ -9,10 +9,6 @@ const upload = multer({ storage });
 
 router.get('/usuario',cadastroController.usuario);
 
-router.get('/pet',cadastroController.pet);
-
 router.post('/usuario',upload.single('fotoPerfil'),cadastroController.novoUsuarios);
-
-router.post('/pet',cadastroController.novoAnimais);
 
 module.exports= router;
