+++ src/views/cadastro-pessoas.ejs
@@ -88,8 +88,8 @@
                     <input type="radio" name="genero" value="mulher"><label for="mulher">Feminino</label>
                 </div>
                 
-                <label for="fotoAnimais"></label>
-                <input type="file" id="fotoAnimais" name="fotoAnimais" required accept="image/*">
+                <label for="fotoPerfil"></label>
+                <input type="file" id="fotoPerfil" name="fotoPerfil" required accept="image/*">
                 
                 <button type="submit"> Finalizar </button>
