+++ usuario.json
@@ -1,2 +1,1 @@
-[
-]
+[{"id":"b454147e-9c78-4814-a21d-b675f112359a","nome":"konrad","cpf":"","email":"brunorepelex@gmail.com","date":""," estado ":" RJ ","cidade":"Cabo Frio","genero":"homem","fileName":"bdac4cb2-c798-4db2-b497-8ca3070fd825-pouting-face-google.png"}]
